made a bunch of changes without regular commits.
I've enabled a cool demo of controlling separate sprites on the same brainpad, except I broke it with what I thought was a smart idea.

I went for handling updates via the messagebus, which sounded like a good idea initially but doesn't scale... It then caused some ugliness with sprite indexing and I had to do a messy translation. Will undo in a few hours.

Better implementation is maintain a reference of sprites for each driver, and iterate every x ms checking if dirty.